refactor: 移动 Gateway Model 到 asset 目录下
