--- apps/assets/models/asset/host.py
@@ -1,6 +1,0 @@
-from assets.const import GATEWAY_NAME
-from .common import Asset
-
-
-class Host(Asset):
-    pass
