+++ apps/assets/models/asset/host/__init__.py
@@ -1,0 +1,2 @@
+from .host import *
+from .gateway import *
