+++ apps/assets/models/asset/host/host.py
@@ -1,0 +1,7 @@
+from ..common import Asset
+
+__all__ = ['Host']
+
+
+class Host(Asset):
+    pass
